+++ src/main/java/micdoodle8/mods/galacticraft/core/client/render/entities/layer/LayerOxygenParachute.java
@@ -66,19 +66,19 @@ public class LayerOxygenParachute implements LayerRenderer<AbstractClientPlayer>
                 {
                     FMLClientHandler.instance().getClient().renderEngine.bindTexture(ModelPlayerGC.playerTexture);
 
-                    this.parachute[0].rotateAngleZ = (float) (30F / Constants.RADIANS_TO_DEGREES);
-                    this.parachute[2].rotateAngleZ = (float) -(30F / Constants.RADIANS_TO_DEGREES);
-                    this.parachuteStrings[0].rotateAngleZ = (float) (155F / Constants.RADIANS_TO_DEGREES);
-                    this.parachuteStrings[0].rotateAngleX = (float) (23F / Constants.RADIANS_TO_DEGREES);
+                    this.parachute[0].rotateAngleZ = 30F / Constants.RADIANS_TO_DEGREES;
+                    this.parachute[2].rotateAngleZ = -(30F / Constants.RADIANS_TO_DEGREES);
+                    this.parachuteStrings[0].rotateAngleZ = 155F / Constants.RADIANS_TO_DEGREES;
+                    this.parachuteStrings[0].rotateAngleX = 23F / Constants.RADIANS_TO_DEGREES;
                     this.parachuteStrings[0].setRotationPoint(-9.0F, -7.0F, 2.0F);
-                    this.parachuteStrings[1].rotateAngleZ = (float) (155F / Constants.RADIANS_TO_DEGREES);
-                    this.parachuteStrings[1].rotateAngleX = (float) -(23F / Constants.RADIANS_TO_DEGREES);
+                    this.parachuteStrings[1].rotateAngleZ = 155F / Constants.RADIANS_TO_DEGREES;
+                    this.parachuteStrings[1].rotateAngleX = -(23F / Constants.RADIANS_TO_DEGREES);
                     this.parachuteStrings[1].setRotationPoint(-9.0F, -7.0F, 2.0F);
-                    this.parachuteStrings[2].rotateAngleZ = (float) -(155F / Constants.RADIANS_TO_DEGREES);
-                    this.parachuteStrings[2].rotateAngleX = (float) (23F / Constants.RADIANS_TO_DEGREES);
+                    this.parachuteStrings[2].rotateAngleZ = -(155F / Constants.RADIANS_TO_DEGREES);
+                    this.parachuteStrings[2].rotateAngleX = 23F / Constants.RADIANS_TO_DEGREES;
                     this.parachuteStrings[2].setRotationPoint(9.0F, -7.0F, 2.0F);
-                    this.parachuteStrings[3].rotateAngleZ = (float) -(155F / Constants.RADIANS_TO_DEGREES);
-                    this.parachuteStrings[3].rotateAngleX = (float) -(23F / Constants.RADIANS_TO_DEGREES);
+                    this.parachuteStrings[3].rotateAngleZ = -(155F / Constants.RADIANS_TO_DEGREES);
+                    this.parachuteStrings[3].rotateAngleX = -(23F / Constants.RADIANS_TO_DEGREES);
                     this.parachuteStrings[3].setRotationPoint(9.0F, -7.0F, 2.0F);
 
                     GlStateManager.pushMatrix();
